Use Yarn in favor of NPM and Bower
This was suggested by Bower, but not all dependencies can be installed
with Yarn (bitters, neat), so we still have to rely on Bower for these
components.

The README was missing the `bower` steps, so these were added.

The FlexSlider fonts were removed, as these are now also provided by
Bower.