Use a Docker container as the build environment
Yarn dependency was removed in the process; use Nodejs instead.

The make-locales script was removed, because you can install
pre-compiled locale data by installing the "locales-all" package.

Running `docker/dev.sh` automatically sets up a build environment and
builds the website.

For now you'll need to use

    docker cp \
    dev_getgnulinux_run_1:/home/ggl/getgnulinux/docroot build

while the build environment is running to create an export of the build.