Use separate Docker container for CI
The volumes of the development container result in permission errors on
Travis CI and I don't know why (maybe because Travis CI uses a different
uid than the one from the development container). The CI container does
not have/need these volumes.

This also removes the entrypoint from the Dockerfile because it breaks
the `run` command from `run.sh`. This wasn't an issue for the
development container because we override the entrypoint there.