Explain "Permission denied" error in the README
Along with an explanation of how to work around this issue.

We can not simply create a new user with the user's current UID because
it might collide with the `node` user created by the base container. One
way to fix that is to install nodejs ourselves, but I prefer not to do
this.