Use GNU Autoconf to generate Makefiles
This makes life easier for me, as I no longer have to maintain a verbose
Makefile for generating .po and .mo files. All I have to do now is keep
the set of available languages in 'po/LINGUAS' up-to-date, and the
Autoconf macro's will do the rest.

This change adds some files generated by `autoreconf --install`, but not
all files. One must run `autoreconf --install` again to create missing
files.

The Makefiles can then be created with:

    ./configure --localedir=/home/node/getgnulinux/docroot/locale/

One can then build and install the translation files using:

    make install-mo

The `make develop` target will do this by default, so there's no need to
execute it manually.

The translation comment prefix was changed from 'translators:' to
'TRANSLATORS:', since this is the default used in the Makefiles from GNU
autoconf. For the same reason the 'locale/getgnulinux' directory was
renamed to 'po'.