Use Serif font for content
And keep using a Sans-Serif font for the rest.

I found the recently added wall-of-text that is the Linux is not Windows
page unpleasant to read with the Open Sans font. That page makes a lot
of use of bold and italic text, and these do not appear clearly with
Open Sans. The default Serif font, however, makes the text much more
legible and is also more pleasant on the eyes.

A Sans-Serif font still looks better for things like buttons and the
navigation items. The default font stack for Materialize looks good
enough to me. I see no reason to load Open Sans for just these elements,
so the Open Sans font was removed completely.