Include the Virus essay by Rick Moen
This article was modified for getgnulinux.org. All instances of "Linux"
were replaced by "GNU/Linux" when referring to the operating system.
External links were marked with `rel="external"`. Abbreviations like
"OS" were replaced by their canonical form.

Also link to it from the Linux FAQ, where an excerpt is included in the
form of an answer.

The article is purposely not made translatable, as it is quite the
wall-of-text.