Deploy via GitLab CI
Add a GitLab CI pipeline for deploying to production. Two piplines are
added: 1) for merge requests and 2) for the master branch. Both
pipelines build a Docker image and run the build scripts. Only the
pipeline for the master branch will deploy to production.

This implements both CI and CD because any commit merged to the master
branch will trigger the build + deploy pipeline. Thus translators no
longer need to wait for manual intervention for their translations to be
deployed.

This replaces Travis CI builds, so the old CI scripts were removed.