Install mo files are part of the `all` target
The `install` target is a little misleasing. This is usually used to
install a software on a computer, but this just installs the mo files to
the docroot directory. So it makes more sense to make this part of the
`all` target, which is used to build the docroot directory.